requirements.txt
awsebcli==3.14.6
blessed==1.15.0
botocore==1.12.3
cached-property==1.5.1
cement==2.8.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
Django==1.11.10
docker==3.5.0
docker-compose==1.21.2
docker-py==1.7.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
misaka==2.1.0
pathspec==0.5.5
Pillow==5.3.0
psycopg2==2.7.5
pycparser==2.19
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.5
PyYAML==3.13
requests==2.18.4
semantic-version==2.5.0
six==1.11.0
sorl-thumbnail==12.5.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.53.0
